#146 editing ratings/tags automatically sets a metadata date via XMP for GIF
